refactor(builder): factory 패키지 이름 변경 -> builder
