+++ src/test/java/kitchenpos/builder/MenuGroupBuilder.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.factory;
+package kitchenpos.builder;
 
 import kitchenpos.domain.MenuGroup;
