+++ src/test/java/kitchenpos/builder/OrderTableBuilder.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.factory;
+package kitchenpos.builder;
 
 import kitchenpos.domain.OrderTable;
