+++ src/test/java/kitchenpos/builder/TableGroupBuilder.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.factory;
+package kitchenpos.builder;
 
 import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.TableGroup;
